Clean univariate time-series data

cleanTS()is the main function of the package which creates a cleanTS object. It performs all the different data cleaning tasks, such as converting the timestamps to proper format, imputation of missing values, handling outliers, etc.

cleanTS(
  data,
  date_format,
  imp_methods = c("na_interpolation", "na_locf", "na_ma", "na_kalman"),
  time = NULL,
  value = NULL,
  replace_outliers = T
)

Arguments

data
A data frame containg the input data. If it contains more than two columns then specify the names of time and value columns using the time and value arguments.

date_format
Format of timestamps used in the data. It uses lubridate formats as mentioned here.

imp_methods
The imputation methods to be used.

time
The name of column in provided data to be used as time column.

value
The name of column in provided data, to be used as value column.

replace_outliers
Boolean, if TRUE then the outliers found will be removed and imputed using the given imputation methods.

Value

A cleanTS object which contains:

Cleaned data

Missing timestamps

Duplicate timestamps

Imputation errors

Outliers

Outlier imputation errors

Examples

if (FALSE) {
  cts <- cleanTS(data = data, date_format = c("dmyHMS"), replace_outliers = F)
}