Merge Multiple CSV files

mergecsv() takes a folder containing CSV files and merges them into a single data.table. It is assumed that the first column of all the CSVs contains the timestamps.

mergecsv(path, formats)

Arguments

path
Path to the folder.

formats
Datetime formats.

Value

Merged data.table.