Check input data

This function is used to check and verify the input data given as input. The package needs a univariate time series as input. This function keeps the first 2 columns, first is renamed as time and second is renamed as value. If the optional time and value arguments are provided then they are used to determine the relevant columns in the data.

check_input(df, dt_format, time, value)

Arguments

df

A data frame containing the input data. If it contains more than two columns then specify the names of time and value columns using the time and value arguments.

dt_format

Format of timestamps used in the data. It uses lubridate formats as mentioned here.

time

The name of column in provided data to be used as time column.

value

The name of column in provided data, to be used as value(observations) column.

Value

Data containing 2 columns, time and value. Time column is converted to POSIX object and value to numeric.