Clean univariate time-series data

cleanTS()is the main function of the package which creates a cleanTS object. It performs all the different data cleaning tasks, such as converting the timestamps to proper format, imputation of missing values, handling outliers, etc.

cleanTS(
  data,
  date_format,
  imp_methods = c("na_interpolation", "na_locf", "na_ma", "na_kalman"),
  time = NULL,
  value = NULL,
  replace_outliers = TRUE
)

Arguments

data
A data frame containing the input data. By default, it considers that the first column to contain the timestamps and the second column contains the observations.If that is not the case or if it contains more than two columns then specify the names of time and value columns using the time and value arguments.

date_format
Format of timestamps used in the data. It uses lubridate formats as mentioned here. More than one formats can be using a vectors of strings.

imp_methods
The imputation methods to be used.

time
Optional, the name of column in provided data to be used as time column.

value
Optional, the name of column in provided data, to be used as value column.

replace_outliers
Boolean, if TRUE then the outliers found will be removed and imputed using the given imputation methods.

Value

A cleanTS object which contains:

Cleaned data

Missing timestamps

Duplicate timestamps

Imputation errors

Outliers

Outlier imputation errors

Examples

# Convert sunspots.month to dataframe
data <- timetk::tk_tbl(sunspot.month)

#> Warning: `type_convert()` only converts columns of type 'character'.
#> - `df` has no columns of type 'character'

print(data)

#> # A tibble: 3,177 × 2
#>    index     value
#>    <yearmon> <dbl>
#>  1 Jan 1749   58  
#>  2 Feb 1749   62.6
#>  3 Mar 1749   70  
#>  4 Apr 1749   55.7
#>  5 May 1749   85  
#>  6 Jun 1749   83.5
#>  7 Jul 1749   94.8
#>  8 Aug 1749   66.3
#>  9 Sep 1749   75.9
#> 10 Oct 1749   75.5
#> # … with 3,167 more rows

# Randomly insert missing values to simulate missing value imputation
set.seed(10)
ind <- sample(nrow(data), 100)
data$value[ind] <- NA

# Perform cleaning
cts <- cleanTS(data, date_format = "my", time = "index", value = "value")

#> Warning: `type_convert()` only converts columns of type 'character'.
#> - `df` has no columns of type 'character'

print(cts)

#> $clean_data
#> # A tibble: 3,177 × 5
#>    time                value missing_type method_used is_outlier
#>    <dttm>              <dbl> <chr>        <chr>       <lgl>     
#>  1 1749-01-01 00:00:00  58   NA           NA          FALSE     
#>  2 1749-02-01 00:00:00  62.6 NA           NA          FALSE     
#>  3 1749-03-01 00:00:00  70   NA           NA          FALSE     
#>  4 1749-04-01 00:00:00  55.7 NA           NA          FALSE     
#>  5 1749-05-01 00:00:00  85   NA           NA          FALSE     
#>  6 1749-06-01 00:00:00  83.5 NA           NA          FALSE     
#>  7 1749-07-01 00:00:00  94.8 NA           NA          FALSE     
#>  8 1749-08-01 00:00:00  66.3 NA           NA          FALSE     
#>  9 1749-09-01 00:00:00  75.9 NA           NA          FALSE     
#> 10 1749-10-01 00:00:00  75.5 NA           NA          FALSE     
#> # … with 3,167 more rows
#> 
#> $missing_ts
#> POSIXct of length 0
#> 
#> $duplicate_ts
#> POSIXct of length 0
#> 
#> $imp_methods
#> [1] "na_interpolation, na_locf, na_ma, na_kalman"
#> 
#> $mcar_err
#> # A tibble: 1 × 4
#>   na_interpolation na_locf na_ma na_kalman
#>              <dbl>   <dbl> <dbl>     <dbl>
#> 1             2.30    2.92  2.25      2.25
#> 
#> $mar_err
#> # A tibble: 1 × 4
#>   na_interpolation na_locf na_ma na_kalman
#>              <dbl>   <dbl> <dbl>     <dbl>
#> 1            0.299   0.407 0.263     0.268
#> 
#> $outliers
#> # A tibble: 39 × 4
#>    time                value orig_value method_used
#>    <dttm>              <dbl>      <dbl> <chr>      
#>  1 1749-11-01 00:00:00  78.4       159. na_ma      
#>  2 1769-09-01 00:00:00 111.        149. na_kalman  
#>  3 1769-10-01 00:00:00 111.        158. na_kalman  
#>  4 1769-11-01 00:00:00 110.        148. na_kalman  
#>  5 1771-05-01 00:00:00  79.9       153. na_ma      
#>  6 1777-12-01 00:00:00 122.        157. na_kalman  
#>  7 1778-01-01 00:00:00 122.        177. na_kalman  
#>  8 1778-05-01 00:00:00 141.        239. na_kalman  
#>  9 1778-06-01 00:00:00 143.        172. na_kalman  
#> 10 1778-09-01 00:00:00 145.        172. na_ma      
#> # … with 29 more rows
#> 
#> $outlier_mcar_err
#> # A tibble: 1 × 4
#>   na_interpolation na_locf na_ma na_kalman
#>              <dbl>   <dbl> <dbl>     <dbl>
#> 1            0.968    1.18 0.956     0.958
#> 
#> $outlier_mar_err
#> # A tibble: 1 × 4
#>   na_interpolation na_locf na_ma na_kalman
#>              <dbl>   <dbl> <dbl>     <dbl>
#> 1             1.38    1.63  1.44      1.26
#>