Clean univariate time-series data

cleanTS()is the main function of the package which creates a cleanTS object. It performs all the different data cleaning tasks, such as converting the timestamps to proper format, imputation of missing values, handling outliers, etc. It is a wrapper function that calls all the other internal functions to performs different data cleaning tasks.

cleanTS(
  data,
  date_format,
  imp_methods = c("na_interpolation", "na_locf", "na_ma", "na_kalman"),
  time = NULL,
  value = NULL,
  replace_outliers = TRUE
)

Arguments

data

A data frame containing the input data. By default, it considers that the first column to contain the timestamps and the second column contains the observations.If that is not the case or if it contains more than two columns then specify the names of time and value columns using the time and value arguments.

date_format

Format of timestamps used in the data. It uses lubridate formats as mentioned here. More than one formats can be using a vectors of strings.

imp_methods

The imputation methods to be used.

time

Optional, the name of column in provided data to be used as time column.

value

Optional, the name of column in provided data, to be used as value column.

replace_outliers

Boolean, if TRUE then the outliers found will be removed and imputed using the given imputation methods.

Value

A cleanTS object which contains:

Cleaned data

Missing timestamps

Duplicate timestamps

Imputation errors

Outliers

Outlier imputation errors

Details

The first task is to check the input time series data for structural and data type-related errors. Since the functions need univariate time series data, the input data is checked for the number of columns. By default, the first column is considered to be the time column, and the second column to be the observations. Alternatively, if the time and value arguments are given, then those columns are used. The time column is converted to a POSIX object. The value column is converted to a numeric type. The column names are also changed to time and value. All the data is converted to a data.table object. This data is then passed to other functions to check for missing and duplicate timestamps. If duplicate timestamps are found, then the observation values are checked. If the observations are the same, then only one copy of that observation is kept. But if the observations are different, then it is not possible to find the correct one, so the observation is set to NA. This data is the passed to a function for finding and handling missing observations. The methods given in the imp_methods argument are compared and selected. The MCAR and MAR values are handled seperately. After the best methods are found, imputation is performed using those methods. The user can also pass user-defined functions for comparison. The user-defined function should follow the structure as the default functions. It should take a numeric vector containing missing values as input, and return a numeric vector of the same length without missing values as output. Once the missing values are handled the data is checked for outliers. If the replace_outliers parameter is set to TRUE in the cleanTS() function, then the outliers are replaced by NA and imputed using the procedure mentioned for imputing missing values. Then it creates a cleanTS object which contains the cleaned data, missing timestamps, duplicate timestamps, imputation methods, MCAR imputation error, MAR imputation error, outliers, and if the outliers are replaced then imputation errors for those imputations are also included. The cleanTS object is returned by the function.

Examples

# Convert sunspots.month to dataframe
data <- timetk::tk_tbl(sunspot.month)
print(data)
#> # A tibble: 3,177 × 2
#>    index     value
#>    <yearmon> <dbl>
#>  1 Jan 1749   58  
#>  2 Feb 1749   62.6
#>  3 Mar 1749   70  
#>  4 Apr 1749   55.7
#>  5 May 1749   85  
#>  6 Jun 1749   83.5
#>  7 Jul 1749   94.8
#>  8 Aug 1749   66.3
#>  9 Sep 1749   75.9
#> 10 Oct 1749   75.5
#> # … with 3,167 more rows

# Randomly insert missing values to simulate missing value imputation
set.seed(10)
ind <- sample(nrow(data), 100)
data$value[ind] <- NA

# Perform cleaning
cts <- cleanTS(data, date_format = "my", time = "index", value = "value")
print(cts)
#> $clean_data
#> # A tibble: 3,177 × 5
#>    time                value missing_type method_used is_outlier
#>    <dttm>              <dbl> <chr>        <chr>       <lgl>     
#>  1 1749-01-01 00:00:00  58   NA           NA          FALSE     
#>  2 1749-02-01 00:00:00  62.6 NA           NA          FALSE     
#>  3 1749-03-01 00:00:00  70   NA           NA          FALSE     
#>  4 1749-04-01 00:00:00  55.7 NA           NA          FALSE     
#>  5 1749-05-01 00:00:00  85   NA           NA          FALSE     
#>  6 1749-06-01 00:00:00  83.5 NA           NA          FALSE     
#>  7 1749-07-01 00:00:00  94.8 NA           NA          FALSE     
#>  8 1749-08-01 00:00:00  66.3 NA           NA          FALSE     
#>  9 1749-09-01 00:00:00  75.9 NA           NA          FALSE     
#> 10 1749-10-01 00:00:00  75.5 NA           NA          FALSE     
#> # … with 3,167 more rows
#> 
#> $missing_ts
#> POSIXct of length 0
#> 
#> $duplicate_ts
#> POSIXct of length 0
#> 
#> $imp_methods
#> [1] "na_interpolation, na_locf, na_ma, na_kalman"
#> 
#> $mcar_err
#> # A tibble: 1 × 4
#>   na_interpolation na_locf na_ma na_kalman
#>              <dbl>   <dbl> <dbl>     <dbl>
#> 1             2.30    2.92  2.25      2.25
#> 
#> $mar_err
#> # A tibble: 1 × 4
#>   na_interpolation na_locf na_ma na_kalman
#>              <dbl>   <dbl> <dbl>     <dbl>
#> 1            0.299   0.407 0.263     0.268
#> 
#> $outliers
#> # A tibble: 39 × 4
#>    time                value orig_value method_used
#>    <dttm>              <dbl>      <dbl> <chr>      
#>  1 1749-11-01 00:00:00  78.4       159. na_ma      
#>  2 1769-09-01 00:00:00 111.        149. na_kalman  
#>  3 1769-10-01 00:00:00 111.        158. na_kalman  
#>  4 1769-11-01 00:00:00 110.        148. na_kalman  
#>  5 1771-05-01 00:00:00  79.9       153. na_ma      
#>  6 1777-12-01 00:00:00 122.        157. na_kalman  
#>  7 1778-01-01 00:00:00 122.        177. na_kalman  
#>  8 1778-05-01 00:00:00 141.        239. na_kalman  
#>  9 1778-06-01 00:00:00 143.        172. na_kalman  
#> 10 1778-09-01 00:00:00 145.        172. na_ma      
#> # … with 29 more rows
#> 
#> $outlier_mcar_err
#> # A tibble: 1 × 4
#>   na_interpolation na_locf na_ma na_kalman
#>              <dbl>   <dbl> <dbl>     <dbl>
#> 1            0.968    1.18 0.956     0.958
#> 
#> $outlier_mar_err
#> # A tibble: 1 × 4
#>   na_interpolation na_locf na_ma na_kalman
#>              <dbl>   <dbl> <dbl>     <dbl>
#> 1             1.38    1.63  1.44      1.26
#>